+++ src/modules/apps/testCommand.ts
@@ -3,7 +3,7 @@ import chalk from 'chalk'
 import { concat, map, prop } from 'ramda'
 import { ManifestEditor } from '../../api'
 import { Builder } from '../../api/clients/IOClients/apps/Builder'
-import { CommandError } from '../../api/error/errors'
+import { createFlowIssueError } from '../../api/error/utils'
 import { createPathToFileObject } from '../../api/files/ProjectFilesManager'
 import { YarnFilesManager } from '../../api/files/YarnFilesManager'
 import log from '../../api/logger'
@@ -125,7 +125,7 @@ export default async options => {
       }
 
       if (data.code === 'link_on_production') {
-        throw new CommandError(
+        throw createFlowIssueError(
           `Please use a dev workspace to test apps. Create one with (${chalk.blue(
             'vtex use <workspace> -rp'
           )}) to be able to test apps`